pyproject.toml
[tool.poetry]
name = "poc-odm-rest-fsm-admin"
version = "0.0.1"
description = "POC for using ODM, REST, FSM and Admin with Flask project."
authors = ["Anup Dhabarde <anup.dhabarde@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3"
flask-mongoengine-3 = "^1.0.9"
Flask-WTF = "^1.2.1"
Flask-Admin = "^1.6.1"
email-validator = "^2.1.1"
flask-restx = "^1.3.0"
marshmallow-mongoengine = "^0.31.2"

[tool.poetry.group.dev.dependencies]
Flask-DebugToolbar = "^0.14.1"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
pytest-flask = "^1.3.0"
mongomock = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py311']

[tool.isort]
profile = "black"
skip = ".cache"
